requirements.txt
pymysql==1.1.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
rich==13.9.4
pandas>=2.2.3